// src/clients/CatalogAdministration/models/PublishingScope.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kibo Catalog Admin Services
 * OpenAPI Spec for Kibo Catalog Admin Services
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PublishingScope
 */
export interface PublishingScope {
    /**
     * True = scope to all pending products
     * @type {boolean}
     * @memberof PublishingScope
     */
    allPending?: boolean | null;
    /**
     * List of specific productCodes to scope publish operation against
     * @type {Array<string>}
     * @memberof PublishingScope
     */
    productCodes?: Array<string> | null;
    /**
     * The Publish Set to publish
     * @type {string}
     * @memberof PublishingScope
     */
    publishSetCode?: string | null;
}

/**
 * Check if a given object implements the PublishingScope interface.
 */
export function instanceOfPublishingScope(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function PublishingScopeFromJSON(json: any): PublishingScope {
    return PublishingScopeFromJSONTyped(json, false);
}

export function PublishingScopeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PublishingScope {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'allPending': !exists(json, 'allPending') ? undefined : json['allPending'],
        'productCodes': !exists(json, 'productCodes') ? undefined : json['productCodes'],
        'publishSetCode': !exists(json, 'publishSetCode') ? undefined : json['publishSetCode'],
    };
}

export function PublishingScopeToJSON(value?: PublishingScope | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'allPending': value.allPending,
        'productCodes': value.productCodes,
        'publishSetCode': value.publishSetCode,
    };
}